% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_api.R
\name{extrair_api}
\alias{extrair_api}
\alias{blocos}
\alias{deputados}
\alias{eventos}
\alias{frentes}
\alias{legislaturas}
\alias{partidos}
\alias{proposicoes}
\alias{referencias}
\alias{votacoes}
\alias{orgaos}
\title{Funções para extrair informação da API}
\usage{
blocos()

deputados()

eventos()

frentes()

legislaturas()

partidos()

proposicoes()

referencias()

votacoes()

orgaos()
}
\description{
A API de dados abertos da camara legislativa possui várias informações dos
deputados, como por exemplo: nome, foto, email, votações, entre outros.
}
\examples{
\dontrun{
deputados()
}

}
